quiz react

+++ app/src/database/migrations/2022_04_24_111400_app_base.php
@@ -30,12 +30,12 @@ return new class extends Migration
         Schema::create('budget_states', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->integer('income_first_year', false, true);
-            $table->integer('income_second_year', false, true);
-            $table->integer('income_third_year', false, true);
-            $table->integer('expense_first_year', false, true);
-            $table->integer('expense_second_year', false, true);
-            $table->integer('expense_third_year', false, true);
+            $table->bigInteger('income_first_year', false, true);
+            $table->bigInteger('income_second_year', false, true);
+            $table->bigInteger('income_third_year', false, true);
+            $table->bigInteger('expense_first_year', false, true);
+            $table->bigInteger('expense_second_year', false, true);
+            $table->bigInteger('expense_third_year', false, true);
             $table->foreignId('budget_chapter_id')
                 ->constrained()
                 ->onUpdate('cascade')
@@ -66,9 +66,9 @@ return new class extends Migration
             $table->timestamps();
             $table->boolean('is_increase');
             $table->boolean('is_expense');
-            $table->integer('first_year', false, true);
-            $table->integer('second_year', false, true);
-            $table->integer('third_year', false, true);
+            $table->bigInteger('first_year', false, true);
+            $table->bigInteger('second_year', false, true);
+            $table->bigInteger('third_year', false, true);
             $table->foreignId('budget_chapter_id')
                 ->constrained()
                 ->onUpdate('cascade')
